app/requirements.txt
flask
Pillow
torch
torchvision
numpy